orderpg_back/requirements.txt
flask
flask-cors
requests
python-dotenv
apscheduler